+++ extensions/iceworks-material-import/src/services/index.ts
@@ -2,5 +2,6 @@ import * as material from '@iceworks/material-service';
 import * as page from '@iceworks/page-service';
 import * as common from '@iceworks/common-service';
 import * as block from '@iceworks/block-service';
+import * as component from '@iceworks/component-service';
 
-export default { material, page, common, block };
+export default { material, page, common, block, component };
