--- packages/block-service/src/utils/generateComponentName.ts
@@ -1,5 +1,0 @@
-import * as upperCamelCase from 'uppercamelcase';
-
-export default function generateComponentName(name: string): string {
-  return upperCamelCase(name);
-}
