--- packages/block-service/src/utils/getImportTemplate.ts
@@ -1,3 +1,0 @@
-export default function getImportTemplate(name: string, source: string): string {
-  return `import ${name} from '${source}';\n`;
-}
