--- packages/block-service/src/utils/getTagTemplate.ts
@@ -1,3 +1,0 @@
-export default function getTagTemplate(name: string): string {
-  return `<${name} /> \n`;
-}
