+++ packages/component-service/README.md
@@ -1,0 +1,6 @@
+# Iceworks Component Service
+
+Iceworks component service for VSCode extension.
+
+> Use these APIs only in VSCode extension, not in Node.js.
+> Because we rely on the APIs of VSCode extension.
