+++ packages/component-service/package.json
@@ -1,0 +1,27 @@
+{
+  "name": "@iceworks/component-service",
+  "version": "0.1.0",
+  "description": "Iceworks component service for VSCode extension.",
+  "files": [
+    "lib"
+  ],
+  "main": "lib/index.js",
+  "typings": "lib/index.d.ts",
+  "dependencies": {
+    "@iceworks/common-service": "^0.1.3",
+    "@iceworks/material-utils": "^0.1.3",
+    "@iceworks/project-service": "^0.1.4",
+    "@typescript-eslint/parser": "^1.9.0",
+    "eslint-plugin-react": "^7.14.2",
+    "fs-extra": "^9.0.1",
+    "typescript": "^3.8.3",
+    "uppercamelcase": "^3.0.0"
+  },
+  "devDependencies": {
+    "@types/vscode": "^1.45.0"
+  },
+  "publishConfig": {
+    "access": "public"
+  },
+  "gitHead": "fc5b35f95ab4cc24898845916acf598c2f34d576"
+}
