+++ packages/component-service/src/utils/checkTemplate.ts
