+++ packages/component-service/src/utils/generateComponentName.ts
@@ -1,0 +1,5 @@
+import * as upperCamelCase from 'uppercamelcase';
+
+export default function generateComponentName(name: string): string {
+  return upperCamelCase(name);
+}
