+++ packages/material-ui/package.json
@@ -54,6 +54,6 @@
     "access": "public"
   },
   "license": "MIT",
-  "homepage": "https://unpkg.com/@iceworks/material-ui@0.1.6/build/index.html",
+  "homepage": "https://unpkg.com/@iceworks/material-ui@0.1.7/build/index.html",
   "gitHead": "fc5b35f95ab4cc24898845916acf598c2f34d576"
 }
